Refactor auth for consistency
