+++ package.json
@@ -158,6 +158,7 @@
       "src/js/base/**/*.js",
       "src/js/auth.js",
       "src/js/auth0.js",
+      "src/js/auth/*.js",
       "src/js/formapp.js",
       "src/js/formapp/**/*.js",
       "src/js/formservice.js",
