+++ src/js/auth/config.js
@@ -1,0 +1,5 @@
+export const PATH_ROOT = '/';
+export const PATH_RWELL = '/rw';
+export const PATH_AUTHD = '/authenticated';
+export const PATH_LOGIN = '/login';
+export const PATH_LOGOUT = '/logout';
